fix(constellation): Change several `Constellation` functions to use `Constellation::root_directory` and `Constellation::root_directory_mut`
This is needed because `root_directory` is meant for the whole index while `current_directory` is meant for the point of reference of the current directory  that is used.

Note: As it stands, `Constellation::current_directory` and `Constellation::current_directory_mut` may be removed in the future since it's no longer has a use at the present time.