chore(data): Added `bincode` into `warp-data` and implemented `Payload`
Payload will be the handler of data for data object, which would allow the use of either `serde_json::Value` or `Vec<u8>`, depending on if the feature selected to enable `bincode`.

Reasoning for this is due to the fact that bincode would give an error `SequenceMustHaveLength`for any object that uses `serde(flatten)`, which due to its design that its not able to supply correct information, but also bincode developers has put it as an "out of scope" issue since bincode only supports types with known lengths. This may change in the future, but for now to keep compatibility this is behind a feature to allow enabling or disabling of the use of bincode with serde_json being its fallback.