chore(tui): Updated `warp-tui`
Implemented layout and interface

Added tui-logger for logging

Note: There is an issue which results in the information to be displayed in the background of the terminal. To work around this, `Terminal::clear` is being called for each update until this is resolved.

TODO: Implement modules and extensions and implement functions for generation of dummy/mock data.