chore(constellation): Removed several functions from `Constellation` and added comments.
Several functions within `Constellation` points back to using `Constellation::root_directory` (or its mutable equivalent) for its functionality, which may become counterproductive due to nested contents. Therefore, they are being removed along with unused functions to allow for one to only need to call a couple of functions for interaction with the filesystem.