feat(constellation): Changed `ConstellationGetPut` to support async.
This change removes the possibility of using a reader or writer as a parameter but instead will use a path directly and have the implementation to implement ways of handling uploading or downloading of the file.

Note: May add additional traits to handle buffers and research a way to incorporate a stream while keeping the trait object safe.