chore(common): Added and re-exported `cfg-if` and `dirs` and other misc changes
Due to bincode being a optional crate and being used under a cfg condition, we would get an error if its not being used but existing within the `Error` enum resulting in us to temporarily have two enums under those conditions. Now, bincode is pushed into its own error handle and having a condition against that to include it if the feature is enabled or to return an empty enum if its not to get rid of the need of duplication or annoying cfg flags for every error.

`cfg-if` crate will also allow us to keep condition clean when checking for features, targets, etc.

Changed targets to reflect towards the `target_family`