chore(fs): Add a default bucket name if none is define
Storj uses a bucket to define where files will be uploaded but some extensions may not have this. To help maintain compatibility, we will check to see if its in the proper format and if a bucket isnt defined in a `bucket://file` format (where "bucket" is the name of the bucket) then `root` will be the default bucket name.