chore(warp): Updated `warp`
Simplified module manager. This will get redesigned to support an array of extensions.

Implemented a basic use of the configuration. If a flag isnt defined to point to the configuration file it will use a default configuration

Added cache to the http (WIP)

TODO: Add export/import structure from cache into constellation and add field into configuration to accept private keys or passwords.