chore(constellation): Updated `warp-constellation`
Change to use `Constellation::root_directory_mut` when importing structure

Added check to place item back into directory if there is an error when moving

Added a recursive field for `Constellation::remove`. This will give some extensions the ability to call for deleting recursively if needed.

Change `Constellation::current_directory` to get the current directory from the internal path. This would remove the need to set or have an external implementation of current path.