chore(tesseract): Switch to use AES-256-GCM and attach nonce to encrypted data
Previous code had it where the key and nonce would be need to be supplied. In this implementation, the nonce will be supplied to the encrypted message itself rather than be carried with the key. This way, the nonce would be unique than for it to be a single nonce and put the whole ciphertext at risk of an attack