chore(tesseract): Generate a SHA256 if key is not 256bit
To allow any form of keys to be used outside of generated 256bit key, `Tesseract::encrypt` and `Tesseract::decrypt` will now check to determine if the key length is 256bit (or 32 bytes). If the key is not, the key will be hashed with SHA256 and the hash will be used as the key for encrypting/decrypting.

Note: It might be a good option to salt the hash as well. Possibly using `nonce` as the salt when hashing to make the password unpredictable and much harder to brute force compared to unsalted hashes.