chore(tesseract): Implemented `Tesseract::import` and `Tesseract::export`
This functions will allow the importing or exporting of a HashMap. If importing, the values of the map will be encrypted before it is stored within Tesseract. When exporting, the contents would be decrypted before a hashmap is returned with the decrypted contents.

Note: Exporting the contents with sensitive data would be done carefully and in a secured environment. If one is not sure that the contents would be safe, then it would be recommended to only save the contents by using `Tesseract::save_to_file`, `Tesseract::save_to_stream`, or any related functions in which the contents would remain encrypted.