chore(solana): Updated `warp-solana-utils`
Changed mnemonic to be a vector in SolanaWallet, depending on the entropy of the mnemonic phrase.

Have it so both mnemonic and keypair are zeroed out when dropped

Added `SolanaWallet::get_pubkey` and `SolanaWallet::get_mnomonic_phrase`

Refract the friends helper. Start base use of the friends program.

Added base of groupchat implementation