warp(crypto): Starting implementation of wasm-bindgen
Changed hashing functions not to rely on `*_hash_stream` due to `Read` not being apart of wasm32

Added `getrandom` and reexported it and uses `getrandom` to for `generate` function

Added `OptionalSalt` for compatibility between a optional slice and a optional vec due to wasm-bindgen not implementing anything for `Option<&[u8]>`. Note: This may get changed to use `Vec<u8>` completely so this would not be required

Note: Due to the way wasm-bindgen handle various of functionality within its crate, it may require using `cfg_if` to have two set of functions, with one handling one set for non-wasm related and another for wasm.

TODO: Finish implementing wasm for other crypto functions.