chore(crypto): Updated `warp-crypto`
Use `Option<Vec<u8>>` instead of `OptionalSalt`

Added `ExtractedData`. This struct is used due to wasm not supporting a tuple variant (eg `(Vec<u8>, Vec<u8>)` ). It is not intended to be used directly but meant to be used instead of the crypto functions for extracting/splitting data.

Made encryption/decryption functions compatible with wasm