chore(constellation): Updated `warp-constellation`
Updated FFI for `Directory` and `Constellation` to have types that point to a pointer.

Use a pointer for `Directory` rather than a raw pointer (eg *mut c_void).

Return a bool for `directory_add_file` and `directory_add_directory` instead of an int. Note: This would

Added exports for `Constellation::select`, `Constellation::export` and `ConstellationDataType`

Added FFI placeholder for `File`, and `Item`