refact(ffi): Change FFI to use trait object indirectly
For some reason when interacting with the pointer in C after the recent layout change, it would cause some functions to segfault, pointing back to the use of the fat pointer (eg *mut Box<dyn T>) when it is dereferenced into a trait object. When using libloading with rust, it works fine so this tells me its related to the use of the pointer of the trait object when using it with C (not tested with any other language yet). To temporarily work around this and also prep for WASM handler for the trait object, it was decided to interact with it indirectly by having a struct hold the trait object and borrow the object.

The change also removed the types that was initially going to be used for compatibility between usage internally and cbindgen.

Note: Internals of FFI may change again