feat(rg): Update warp-rg-libp2p initial implementation
Note: This current implementation relies on a single connection to a topic that the peer subscribed to, rather that be local or through a relay/node or directly to a peer ip (not tested in this implementation). This implementation is also a starting point and will be subjected to a breaking change as the code base is refactored.