add(constellation): Added `Constellation::put_stream` and `Constellation::get_stream`
These functions will allow one to pass a boxed `Stream` from or to these functions to stream data for uploading or downloading without buffering. Can be extremely useful if one with to stream the data from the filesystem to another stream (eg passing the data to another destination such as HTTP, etc) or if one wish to handle saving the stream in their own way without buffering or saving the file directly.