feat(tesseract): Perform a check against contents with provided key
This functionality will perform a check against the content with the private key when unlocking the data store and will return an error if all contents cannot be unlocked with key that is provided. This functionality is used when loading the data store from file or from a reader at the time being and will expand. Check can be disabled with `Tesseract::disable_key_check` prior to unlocking the store.