fix(solana): Check all request when triggering a hook for denying or closing.
Since the incoming and outgoing request go based on the pending status, the hooks would not get triggered since the request that is denied or closed would not be on such list. In this case we would need to look at all request and filter it based on the conditions provided.