chore(solana): Updated warp-mp-solana
Add public key to extra_one field in the user program. This will allow us to work around being unable to get the public key of the user when searching by username.

Implemented search by username.

TODO: Determine what is causing the user program to error when setting any data in one of the extra fields.