chore(ffi): Implemented macro to generate ffi array function
In C, it is not getting the next element in the array causing a segfault. This would have to be investigated so in the mean time, this workaround would allow us to just to create functions for data types that holds an array of this data and allow us to pass through `Vec::get` and `Vec::len`.