chore(rg): Reuse `process_message_event` within trait functions
To reduce duplication of code it was decided to utilize `process_message_event` within the trait function after sending the event off to tokio to be sent over libp2p. Since the function contains the same code as used within the trait functions, it would be better off to reuse the function.

TODO: Implement caching within the function to cache when messages are sent and received.