chore(tesseract): Updated tesseract module
Remove `TesseractShared`

Implement `Arc<Mutex<T>>` internally

Implement a drop counter. This would allow tesseract to keep it unlocked across threads unless it is locked manually. The counter increases with every clone of tesseract and after it is dropped, the counter decreases. If it hits 0, where the last instance remaining is dropped, it will check to see if its unlocked and if it is it will proceed to lock it.

TODO: Look into AtomicUsize and determine if this would be a better choice over `Arc<Mutex<usize>>`.