chore(rg): Update warp-rg-libp2p
Implement a peer registry. This registry will allow for storage of peers and their responding public key that are meant to be compatible with this extension.

Implement bootstrap

Remove injected events for our own "out_event". This would give finger tune of control over the swarm events and allow for async calls when needed.

Implement "FindPeer". This would have the client walk the DHT for the peer. Should be called once bootstrapping is complete.

TODO: Implement bootstrap completion check.