refactor: Relocate PublicKey from multipass to crypto
Due to PublicKey not being specific to identity, this change, although not a breaking change, will relocate PublicKey to the crypto module.

Along with this change will introduce a `PrivateKey` which will be used to store the private key instead of using a vector of bytes. This may change in the future to use a slice of bytes within a enum to represent different key types (eg ed25519, x25519, aes, etc) but for now will use a vector of bytes to represent the various of sizes. Using this instead of bytes will allow us to zero out the the key securely from memory when it drop.