chore(crypto): Update crypto module
Added `Cipher::direct_encrypt` and `Cipher::direct_decrypt`. These will allow for direct encryption and decryption without needing to create an instance of `Cipher`. Useful in a instance where single instance is needed and not where the private key would need to be retained in memory in its current state.