chore(rg): Added a bootstrap_complete flag
Adding this flag and implementing it so it will be marked true when bootstrap is complete will allow us to know when it would be the correct time to query against DHT rather than when libp2p is still bootstrapping. This will also be used as an indicator when debugging any performance degradation.