refactor(mp): Added signature field to `FriendRequest`
This would allow the request to be signed by the sender, which would allow the receiver verify the request. The request can be verify by the restructuring the the struct without the signature field and depending on the extension could serialize it to a json or binary format and use that data to verify the signature using the `from` public key.