chore(gh): Readd cache and add update command
The caching was previously removed due to it using an out dated cache when running a test and would cause a error. After further testing it looks like running `cargo update` after loading the cache would work in updating anything in the cache, however we would still need to be cautious. Since `Cargo.lock` is not committed to github, we should not have to worry to much in regards to the dependencies but still be on the look out for conflicts which may require the `Cargo.lock` to be removed from the cache, if it is cached.