refactor(mp): Have `Friends::list_friends` return `Vec<PublicKey>`
This change is mostly due to the fact that public keys would reference the identity itself and in some cases (eg libp2p or ipfs) where a public key cannot be resolved to an Identity for some reason, it would likely return an error in this function. As a result, this would push one to call `MultiPass::get_identity` while using the public key as the reference for the lookup.