chore(mp): Move task into individual functions
Initially, the friend request management would have sent the request to a separate tokio task to handle the request and return the results over a oneshot sender (which would have included some further interaction down the road with), however since things have changed a bit, the functionality of the request was moved into the corresponding functions. This may get reverted in the future however.