chore(mp): Add example
Note: This uses delays between some calls to give the background task time to catch up (since its not awaiting on the task to complete in the background such as sending a and processing of a request). This example may change in the future but actual use may not require such delays.