refactor(mp): Use specific types
rust-ipfs has two stores called `Types` (for persistence) and `TestTypes` (in-memory/temporary). By default, the current implementation uses `Types` and would erase it if the ipfs block store was meant to be temporary. However, there are cases where we wouldnt want to write to disk, especially in case of tests. This implementation would allow us to give both as a option to use instead of restricting it to just one, which would also enable faster tests.

Note: In case of persistence, it would be best to have tesseract also save to disk (via autosave), otherwise there could be data loss.