fix(rg): Use correct topic for new conversation and store as bytes
Sata encodes using ipld internally which makes use of serde for the codec, but because of it utilizing ipld, it does not serialize/deserialize in the same manner as serde_json, which makes it hard to use enums as data. To get around this, the enum is also converted into bytes and deserialized back into its respected object after decrypting.