chore(rg): Added placeholder for storage and signing
Signing will be used to assure that the message, when being edited, is validated to prevent third party tampering of the message (eg another peer editing the recipient message).

Storage is for messages sent/received (for now direct message but will expand to group messages in the future) that is to sit outside of caching.