chore(rg): Save queue if path is set and added queue_event function
When there is a change in queue, we should automatically save it to disk if its available. With this function, internal functions that pushes a item to use will also save it to disk if the path is available, however because saving it to queue doesnt produce a error, we should not return an error but instead log it if saving it to disk returns an error.