chore(rg): Updated warp-rg-ipfs
Subscribe to topic as conversation loads in

Queue event when creating conversation and peer is not connected

Remove topic field from DirectConversation. This will be referenced by `DirectConversation::topic` which would hold the topic that gossipsub would be subscribed too. This would allow us to change the topic later on without breaking change (eg if we want to change the prefix, etc)