Warp 120 - Direct Messaging Implementation (#9)
* add(rg): Base implementation of direct messaging

This still requires alot more functionality before it can be used but is the beginning of a starting point

* chore(rg): Additional implementation to messaging store

* fix(rg): Use correct return

Note: Internal functionality should begin to use `Result<T, E=warp::error::Error>`

* chore(rg): Started implementation of task

This task would process the initial/main topic messages for creation and deletion of conversations and would eventually pass stream off to another task for processing messages (or its events)

Also implemented `DirectConversation::new_with_id` to create a conversation with a uuid that is received or from storage (when loading conversation).

* chore(rg): Implemented send_event

This will send the direct conversation event

* fix(rg): Use correct topic for new conversation and store as bytes

Sata encodes using ipld internally which makes use of serde for the codec, but because of it utilizing ipld, it does not serialize/deserialize in the same manner as serde_json, which makes it hard to use enums as data. To get around this, the enum is also converted into bytes and deserialized back into its respected object after decrypting.

* chore(rg0: Have list_conversations be async

* chore(rg): Reenable check to determine if conversation exist

* chore(rg): Check queue for connected peers

* feat(rg): Functionality to delete conversation

Internally there is a argument to allow broadcasting it so the other party would also have the conversation deleted. This would likely be a configurable option

* chore(rg): Change the explicit lock

After reviewing the code, the write lock would not be needed as events come in as a "first-come-first-serve" and not processing each event in its own task which would then require explicit locks to prevent data races.

* add(rg): Added an example

* chore(rg): Added message to example

* chore(rg): Remove item out of queue

When a peer is connected to a topic, it will send the message and clear the item out of queue

Note: Doing this assumes they will receive the message if there is no error while sending it, but we should probably take into consideration that even if the peer is shown to be connected that they may not receive the message for some reason (eg poor connectivity to peer, etc).

* chore(rg): Check if conversation exist

Although the conversation is checked later in the function it is best to check a head of time

* chore(rg): Added placeholder for storage and signing

Signing will be used to assure that the message, when being edited, is validated to prevent third party tampering of the message (eg another peer editing the recipient message).

Storage is for messages sent/received (for now direct message but will expand to group messages in the future) that is to sit outside of caching.

* feat(rg): Implement storage for message

* chore(rg): Remove map holding streams

* fix(rg): Use correct path when deleting conversation

* chore(rg): Add a persistent example

This example would allow one to keep data across example load. The data would be stored in a specific directly set by the `--path` flag.

* chore(rg): Add type check

We do not want to save any messages if the type if temporary so with this check, we would make the path = None if T is `Temporary`, otherwise we would continue to use that path.

* chore(rg): Save queue if path is set and added queue_event function

When there is a change in queue, we should automatically save it to disk if its available. With this function, internal functions that pushes a item to use will also save it to disk if the path is available, however because saving it to queue doesnt produce a error, we should not return an error but instead log it if saving it to disk returns an error.

* chore(rg): Updated warp-rg-ipfs

Subscribe to topic as conversation loads in

Queue event when creating conversation and peer is not connected

Remove topic field from DirectConversation. This will be referenced by `DirectConversation::topic` which would hold the topic that gossipsub would be subscribed too. This would allow us to change the topic later on without breaking change (eg if we want to change the prefix, etc)

* chore(rg): Updated example

* chore(rg): Updated examples

Move list_conversation behind an interval. This would make sure that the loop does not hit the RwLock since this command would always return `Result::Ok`.

* chore(rg): Updated example message

* chore(rg): Added example

Note that this uses solana for identity while the communication still uses ipfs

* fix(rg): Add interval

* chore(rg): Updated example

Use flatfile cache for persistent example

* chore(rg): Move DID behind Arc

We use Arc<T> because cloning DID would strip the private key from DID but also want to reduce the need of calling the private key function from multipass multiple times.

* chore(rg): Change error to Error::MessageNotFound

* chore(rg): Boxed DID in enum variant

* chore(rg): Updated warp-rg-ipfs

Added code for topic discovery. This will run in a separate tokio task to find and connect to peers providing the same cid (which represents the hash of the topic) over DHT.

* chore(rg): Added ffi functions

* chore(rg): Check block list before creating or accepting messages

Although being blocked will likely block the peer connection, there is a chance that it would go through since we arent blocking the peer at libp2p initially in which case we should still check the list. This will also prevent new conversations if the peer is blocked.

* chore(misc): Updated Cargo.toml

This will exclude the libp2p extension. It will soon be removed but is left as a reference

* chore(misc): Updated interface

Added create_conversation - Used to create the initial conversation with a user

Added list_conversation - Used to list all conversations.

Modified delete to mark `message_id` as optional. If this is `Option::None` the conversation would get deleted. If an ID is defined the message itself would get deleted.

* fix(ffi): Use mutable pointer instead of casting it to a const pointer

* chore(rg): Rename the enum variants

This is to have clippy stop complaining about the variant containing the same name as the enum prefix

* chore(rg): Enable header generation

* chore(c): Updated C example

* refactor(misc): Add configuration and change configuration layout

With this change would allow for relay to handle a array of addresses rather than just a single optional relay. Relay would also default to using the bootstrap relay

Updated example to enable relay and dcutr by default.

* fix(rg): Use correct value and Option::None instead of returning an error

* fix(rg): Use configuration for interval and discovery