chore(mp): Updated warp-mp-ipfs
Dedup cid from tesseract. Due to rust-ipfs lacking ipns at the moment, the cid of the identity was originally stored in tesseract to have as a reference for us to fetch the identity from ipfs. However, this was only temporary and for now the cid will be stored in its own file at the root of a defined path, if one is set. In the near future this will migrate over to using ipns when its implemented into rust-ipfs to make it easier to fetch the cid from DHT, leaving pubsub as a fallback to broadcast or look for identities.