fix(mp): write to the lock rather than creating a new one
Due to creating a new lock with a reference counter (arc<rwlock<T>>) it caused the a race condition where one instance of the store was handling a blank list while the other was handling the real list. Upon saving it would be attempting to write 2 times. This was fixed in a local commit but was forgotten to place back after a revert of some changes. This also placed in a local atomic counter in the store so when it drops, it only saves one time rather than twice.