WARP-145 - Optimizations  (#12)
* chore(misc): Update configuration

This would allow us to skip serialization if one of the conditions in the macro exist for that field.

* chore(mp): Optimize friends store

This would handle request differently by pushing it into a internal profile while also implementing a queue feature. This still requires additional cleanup but this change would also allow us to properly store it as a dag in the near future and be more consistent

* chore(mp): Additional cleanup

* chore(misc): Small cleanup

* chore(tesseract): Have it enable autosave after loading successfully from a file

* fix(rg): Truncate file when saving

* chore(mp): Use configuration path

* chore(mp): Update example

The example would have a option to store information on disk using the `--path` option.

* chore(mp): Updated path for friends store

* chore(rg): Save queue and change layout for of saved data

This also made it so when broadcasting the request it can be saved or not. Useful for when rejecting a request but do not want to save it internally and other misc changes

* chore(mp): Updated warp-mp-ipfs

Dedup cid from tesseract. Due to rust-ipfs lacking ipns at the moment, the cid of the identity was originally stored in tesseract to have as a reference for us to fetch the identity from ipfs. However, this was only temporary and for now the cid will be stored in its own file at the root of a defined path, if one is set. In the near future this will migrate over to using ipns when its implemented into rust-ipfs to make it easier to fetch the cid from DHT, leaving pubsub as a fallback to broadcast or look for identities.

* chore(mp): Clean up with the request

* chore(misc): check handle before using static

* chore(mp): Added close command in example

* chore(mp): Implement close function

* fix(mp): write to the lock rather than creating a new one

Due to creating a new lock with a reference counter (arc<rwlock<T>>) it caused the a race condition where one instance of the store was handling a blank list while the other was handling the real list. Upon saving it would be attempting to write 2 times. This was fixed in a local commit but was forgotten to place back after a revert of some changes. This also placed in a local atomic counter in the store so when it drops, it only saves one time rather than twice.

* chore(mp): Split the saving up into separate functions

This would allow for saving to specific files rather than saving everything.

* chore(mp): Updated warp-mp-ipfs

Cache identity after fetching.

Note: This may get pushed internally in the store

* fix(mp): Use correct query field

* chore(mp): Update warp-mp-ipfs

Remove cache from the store

* chore(rg): Added example

This example will soon replace the normal and persistent examples since this would be using the trait object direct, which would allow one to enter a password in for tesseract as well as defining a path if they wish for anything to be saved

* chore(mp): Remove extra variable and add small delay in discovery

* chore(misc): Implemented function

Implemented mnemonic_into_tesseract. This will generate the keypair from a mnemonic phrase and store it into tesseract. Keypair that it will generate will be ed25516.

* chore(misc): Update configuration

* add(ffi): Added FFI functions

These functions would export the configuration to warp-mp-ipfs and warp-rg-ipfs

* fix(rg): Changed function names for ffi

* chore(misc): Updated Cargo.toml

Signed-off-by: Darius <darius.clark93@gmail.com>

* chore(misc): Updated Cargo.toml

Signed-off-by: Darius <darius.clark93@gmail.com>

* chore(misc): Change crates

This change will allow warp-fs-ipfs to use rustls.

* chore(c): Updated C example

* chore(mp): Updated example

* chore(mp): Minor cleanup

* chore(mp): Deref read guard

Signed-off-by: Darius <darius.clark93@gmail.com>

Signed-off-by: Darius <darius.clark93@gmail.com>