chore(mp): Added LookupBy::ShortId
This will check against the short id of the identities that is derived from the did key.

Note: This should not be a reliable way to get a single identity as there is a chance of 2 identities having the same slice from the did key. Although the change of it is slim, there is still that chance unless the SHORT_ID_SIZE is increased.

TODO: Perform a check against did key when finding identities and filter it out when it does not match the did of the identity.