fix(mp): Fix deadlock
Deadlock is caused by the internal cache still holding a read lock instead of dropping before writing to the field causing a deadlock on the write lock. Splitting it would ensures that the read lock would drop.

TODO: Replace Vec with HashSet to use HashSet::replace to insert/set or replace data.