Warp 147 (#14)
* refactor(mp): Change short_id to use a slice of the public key

* chore(ffi): Add additional ffi codegen

This will allow for generation of `FFIResult_FFIVec_<T>` along side other `FFIVec` implementation

* refactor(mp): Change `MultiPass::get_identity` to return a vector of identities.

* chore(mp): minor changes

* chore(mp): Make lookup non case-sensitive

* chore(mp): Have a const of the slice size

* chore(mp): Added LookupBy::ShortId

This will check against the short id of the identities that is derived from the did key.

Note: This should not be a reliable way to get a single identity as there is a chance of 2 identities having the same slice from the did key. Although the change of it is slim, there is still that chance unless the SHORT_ID_SIZE is increased.

TODO: Perform a check against did key when finding identities and filter it out when it does not match the did of the identity.

* add(mp): Add additional field to `Identity`

This commit introduce a signature field to `Identity` that would allow for the identity to be signed

* fix(mp): Fix deadlock

Deadlock is caused by the internal cache still holding a read lock instead of dropping before writing to the field causing a deadlock on the write lock. Splitting it would ensures that the read lock would drop.

TODO: Replace Vec with HashSet to use HashSet::replace to insert/set or replace data.