chore(mp): Added small delay to prevent writer starvation
Since we are polling/looping while checking to determine if tesseract is unlocked, we are creating the possibility of writer starvation. Adding a small delay in the loop can give the lock a chance to breathe between each check until it is unlocked but also prevents a lockup when performing any write operations.