feat(rg): Updated warp-rg-ipfs
Implemented initialization to help relax the extension account requirement. This means that you will still need to pass an account but will not perform hard checks unless one own account is available. This will, in theory, allow the extension to be loaded along side warp-mp-ipfs before the account is created.