chore(rg): Add a check against the sender before performing task in `ConversationEvents::DeleteConversation`
Precautionary measure to make sure that one cannot forge an event to send to a user that would cause them to delete the conversation. Its a simple check, but may not be something desirable long term. Long term solution might be to have the direct conversation signed and validated by both parties so one or the other can validate themselves prior to the conversation being deleted (or atleast any attempts)