Reuses passphrase field in MultiPass::create_identity (#18)
* feat(mp): Push initialization into its own function

This change to the initialization of the extension will begin to relaxed the extension so tesseract does not have to be unlocked initially.

Note: In this commit, if keypair doesnt exist, unlocking before creating account is fine, however if keypair exist, it is preferred to have it unlocked prior to initializing the extension

* chore(mp): Implement check against tesseract

This check will allow us to determine if tesseract is unlocked and if it is to perform initialization check to see if keypair is stored in it.

* feat: Use `passphrase` in `MultiPass::create_identity` to accept a mnemonic phrase.

This will allow one to pass a valid mnemonic phrase when creating an identity. If one is not provided a keypair will be generated automatically. If a keypair exist in tesseract already, the phrase will be ignored.

* fix(mp): Remove Drop impl and unused configuration option

* fix(mp): Check to determine if key is generated

This may be due to the keypair being available from a phrase being used to derive the keypair from.

* chore(mp): Added small delay to prevent writer starvation

Since we are polling/looping while checking to determine if tesseract is unlocked, we are creating the possibility of writer starvation. Adding a small delay in the loop can give the lock a chance to breathe between each check until it is unlocked but also prevents a lockup when performing any write operations.

* feat(rg): Updated warp-rg-ipfs

Implemented initialization to help relax the extension account requirement. This means that you will still need to pass an account but will not perform hard checks unless one own account is available. This will, in theory, allow the extension to be loaded along side warp-mp-ipfs before the account is created.

* chore(rg): Remove pub identifier