chore(misc): Implemented additional logging and added a check
Implemented a seen check to determine if peers have already seen the identity and if so to skip sending, otherwise update the internal list and send the identity out. Also has it so when the identity is updated that this list is cleared out.

Note: This sends out to everybody if the list change since we arent able to obtain the public key just yet.

TODO: Implement option to send identity out to specific peers rather than to the whole topic (although this would require upstream changes)