chore: Updated extensions and configuration (#20)
* refactor(config): Update configuration

Remove outdated bootstrap (could still be used but its used by libp2p/ipfs for older protocols)

Added additional configuration options for warp-mp-ipfs and warp-rg-ipfs
* sync - placeholder for servers that would be used to sync identity information
* with_friends - to allow to conversations to be created with friends only
* allow_unsigned_message - to send a message without it being signed. Useful for verification. Note, if the recipient this true, messages will be rejected
* store_decrypted - To store messages in a decrypted format. This is used mostly for debugging purposes.
* discovery_name - a topic to use if one wish to have a separate topic to discover peers. Note: If this is set, you will not be able to discover others thats using the default topic as this would override that

* chore(warp): Updated warp, warp-mp-ipfs and warp-rg-ipfs

Added and reexported tracing

Added basic logging to raygun and multipass extension

Updated ipfs to a difference branch which uses libp2p-next-usable

* chore(misc): Update dependency

* chore(misc): Added a feature to dependency

* chore(mp): Remove logger from examples

* chore(misc): Implemented additional logging and added a check

Implemented a seen check to determine if peers have already seen the identity and if so to skip sending, otherwise update the internal list and send the identity out. Also has it so when the identity is updated that this list is cleared out.

Note: This sends out to everybody if the list change since we arent able to obtain the public key just yet.

TODO: Implement option to send identity out to specific peers rather than to the whole topic (although this would require upstream changes)

* chore(gh): Add protobuf to github action

* chore(rg): Disable discovery if ipfs exist

Because warp-mp-ipfs has discovery as a option, peers will already be connected so we would assume that if warp-mp-ipfs is used and multipass passes into raygun with the handle that the peers will be discovered and connected through the initial discovery and wouldnt need to do a second pass on a different topic.

* chore(mp): Add small placeholder for future functionality

* chore(mp): Added tag to PayloadEvent

* chore(misc): Added tracing-appender

* chore(misc): Use reexported dependency

* chore(ffi): Exclude struct from cbindgen

* add: Implemented logging functions into FFI

This would allow logging to a file or to stdout.

Note: These functions only need to be called once with modules being added in the directive.

* chore(misc): Reexported tracing_subscriber

* chore(misc): added additional flags to configuration

This flag will be used for enabling debug for logging and placeholder for future relays.

* chore(mp): Added swarm configuration

This will begin to allow us to modify libp2p swarm and limits, which would reduce the load it would probably have and possibly resolve some issues related related to file desc (especially when the system has a low set)

* chore(mp): Added logging to identity-interface

Use `RUST_LOG[=CRATE]=LOG_LEVEL where CRATE would be optional but for specific libraries using logging, and LOG_LEVEL for info, debug, error, or trace.

* chore(rg): Added logging for ipfs_messenging example

Same as commit message from 969e45887191ffe9a24ee8031970b5d195467f0b

* chore(mp): Made adjustments to the swarm configuration

* chore(mp): Increase the connection limits

* chore(rg): Move warp-rg-libp2p into deprecated directory

This will get removed in the near future.

* chore(mp): Disable bootstrapping if none is in the list

* add(mp): Added additional configuration for swarm

* chore(mp): Implement config against libp2p swarm config.

* chore(mp): implement Default for ConnectionLimit and Swarm

* chore(misc): Added fdlimit as a development dependency

This will help increase the soft limit on linux and mac on launch

* chore(rg): Yield after spawning task

* chore: Remove some of the logic in topic discovery

* chore(rg): Added fdlimit to examples

* chore: Change from tmp directory to current directory for logging

* chore(mp): Added transport config

* chore(mp): Correct info log message

* chore(mp): Added fdlimit to example

* chore(rg): Updated examples

Added logging and have `get_username` return the DID if a username cannot be found

Note: If the username cannot be found, it is likely an issue somewhere with broadcasting the identity.